[feat] todo 앱 타입 분리
todo앱에 타입을 분리함
