+++ src/index.tsx
@@ -12,7 +12,7 @@ const store = configureStore({
     reducer: {
         rootReducer
     }
-})
+});
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
