+++ src/types/todo.ts
@@ -1,0 +1,39 @@
+export type TodoTemplateProps = {
+    input: string;
+    todos: Todo[];
+    onChangeInput: OnChangeInputType;
+    onInsert: OnInsertType;
+    onRemove: OnRemoveType;
+    onToggle: OnToggleType;
+};
+
+export type TodoListProps = {
+    todos: Todo[];
+    onRemove: OnRemoveType;
+    onToggle: OnToggleType;
+};
+
+export type TodoItemProps = {
+    id: number;
+    text: string;
+    checked: boolean;
+    onRemove: OnRemoveType;
+    onToggle: OnToggleType;
+};
+
+export type TodoInsertProps = {
+    input: string;
+    onChangeInput: OnChangeInputType;
+    onInsert: OnInsertType;
+};
+
+export type OnChangeInputType = (e: string) => void;
+export type OnInsertType = (e: string) => void;
+export type OnRemoveType = (e: number) => void;
+export type OnToggleType = (e: number) => void;
+
+export interface Todo {
+    id: number;
+    text: string;
+    checked: boolean;
+}
